created separate component for fish, location and tips, each with their corresponding .js file
