+++ index.html
@@ -15,9 +15,7 @@
             <a class="navigation__link" href="locations.html">Harvesting Locations</a>
         </nav>
     </header>
-
     <main class="container"> 
-
         <div class="fishList">
             
         </div>
@@ -30,7 +28,7 @@
             <div class="tip">Make sure to feed them daily</div>
         </aside>
     </main>
-    <script type="module" src="./scripts/main.js"></script>
+    <script type="module" src="./scripts/location/main.js"></script>
 </body>
 
 </html>
