+++ main.css
@@ -15,8 +15,7 @@ a{
     background-color: rgb(92, 146, 167);
     width: 100%;
     height: 60px;
-    margin-left: 0;
-    margin-top: 0;
+    margin-top: 20px;
     text-align: center;
     padding: 10px;
     text-decoration: none;
@@ -24,13 +23,13 @@ a{
 }
 
 .navigation__link {
-    padding: 0 1em 0 1em;
+    padding: 0 2em 0 2em;
     color: rgb(255, 255, 255);
 }
 .navigation__link:hover{
-    transform: translate(2px, 50px);
-    transition-delay: 0.5s;
-    border-bottom: 2px solid rgb(236, 184, 115);
+    transform: translate(2px, 5px);
+    transition-delay: 0.1s;
+    color: rgb(247, 202, 4);
 }
 
 .fishList {
