+++ scripts/fish/fishList.js
@@ -1,5 +1,5 @@
-import { useFish } from "./FishDataProvider.js";
-import Fish from "./Fish.js";
+import { useFish } from "../fish/FishDataProvider.js";
+import Fish from "../fish/Fish.js";
 
 const contentTarget = document.querySelector(".fishList")
